use openrouter format with cline api
